scrapper: insert fake OCI digests to unstuck nodes
A bug introduced in coreos-assembler[1] (reverted [2]) ended up creating
disk images that had different OCI digests than the released OCI image.
This results in deployed nodes being unable to update [3] because
Zincati look for the deployed OCI checksum in the graph to determines
the update path [4].

This introduces a workaround for it : one day a week we will change the
image pullspec in the OCI graph with the invalid values to give a
chance to Zincati to trigger an update. This will get the node back to
a valid state.
Note that we won't update the last release in the graph, so Zincati have
a valid checksum to fetch from the registry.

[1] https://github.com/coreos/coreos-assembler/commit/9190a34d56d43c5d17e1cbbde9ef2405f76abcff
[2] https://github.com/coreos/coreos-assembler/pull/4374
[3] https://github.com/coreos/fedora-coreos-tracker/issues/2066
[4] https://github.com/coreos/zincati/blob/238a79a9c2d11a39d7b7f9c6e71888b75d2c6ab3/src/cincinnati/mod.rs#L230-L245